fix: use Integer for entity IDs to match SERIAL column type
PostgreSQL SERIAL maps to INTEGER, not BIGINT. Hibernate validate
mode caught the type mismatch at startup.

+++ spring-boot-app/src/main/java/com/workshop/employee/repository/ApiClientRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Repository
-public interface ApiClientRepository extends JpaRepository<ApiClient, Long> {
+public interface ApiClientRepository extends JpaRepository<ApiClient, Integer> {
 
     Optional<ApiClient> findByClientIdAndClientSecret(String clientId, String clientSecret);
